backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.10
python-dotenv==1.0.0
pydantic==2.5.0
mangum==0.17.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
